Extract response processing to a separate class
